+++ tests/test_field_audit.py
@@ -7,14 +7,16 @@ from django.test import TestCase
 
 from field_audit.field_audit import (
     AlreadyAudited,
+    InvalidManagerError,
     _audited_models,
+    _verify_auditing_manager,
     _decorate_db_write,
     audit_fields,
     get_audited_class_path,
     get_audited_models,
     request as audit_request,
 )
-from field_audit.models import AuditEvent
+from field_audit.models import AuditEvent, AuditingManager
 
 from .models import (
     Aerodrome,
@@ -22,6 +24,7 @@ from .models import (
     CrewMember,
     Flight,
     SimpleModel,
+    ModelWithAuditingManager,
 )
 
 
@@ -46,6 +49,12 @@ class TestFieldAudit(TestCase):
             @audit_fields("field")
             class Test:
                 pass
+
+    def test__verify_auditing_manager_with_incorrect_manager_raises(self):
+        class Item0(models.Model):
+            pass
+        with self.assertRaises(InvalidManagerError):
+            _verify_auditing_manager(Item0)
 
     def test__decorate_db_write_for_invalid_func_raises(self):
         def invalid(self):
@@ -88,9 +97,33 @@ class TestFieldAudit(TestCase):
                 item.unsupported()
                 classmeth.assert_not_called()
 
+    def test_audit_fields_verifies_manager_for_audit_special_queryset_writes(self):  # noqa: E501
+
+        class Item1(models.Model):
+            value = models.IntegerField()
+
+        class Item2(models.Model):
+            value = models.IntegerField()
+            objects = AuditingManager()
+
+        with override_audited_models():
+            audit_fields("value")(Item1)  # doesn't raise
+        with self.assertRaises(InvalidManagerError):
+            audit_fields("value", audit_special_queryset_writes=True)(Item1)
+        with override_audited_models():
+            # doesn't raise
+            audit_fields("value", audit_special_queryset_writes=True)(Item2)
+
     def test_get_audited_models(self):
         self.assertEqual(
-            {Aerodrome, Aircraft, CrewMember, Flight, SimpleModel},
+            {
+                Aerodrome,
+                Aircraft,
+                CrewMember,
+                Flight,
+                SimpleModel,
+                ModelWithAuditingManager,
+            },
             set(get_audited_models()),
         )